[#25] 기존의 ChooseVoiceFragment 코드 수정
